SIG: Move go/sig/disp to go/sig/internal/disp
