+++ go/sig/base/BUILD.bazel
@@ -15,7 +15,7 @@ go_library(
         "//go/lib/infra:go_default_library",
         "//go/lib/log:go_default_library",
         "//go/lib/snet:go_default_library",
-        "//go/sig/disp:go_default_library",
+        "//go/sig/internal/disp:go_default_library",
         "//go/sig/mgmt:go_default_library",
         "//go/sig/sigcmn:go_default_library",
     ],
