+++ go/sig/egress/session/BUILD.bazel
@@ -20,10 +20,10 @@ go_library(
         "//go/lib/snet:go_default_library",
         "//go/lib/spath:go_default_library",
         "//go/lib/spath/spathmeta:go_default_library",
-        "//go/sig/disp:go_default_library",
         "//go/sig/egress/iface:go_default_library",
         "//go/sig/egress/siginfo:go_default_library",
         "//go/sig/egress/worker:go_default_library",
+        "//go/sig/internal/disp:go_default_library",
         "//go/sig/metrics:go_default_library",
         "//go/sig/mgmt:go_default_library",
         "//go/sig/sigcmn:go_default_library",
