+++ go/sig/internal/disp/disp.go
